added controlls
